distributed-training/nlp/requirements.txt
--extra-index-url https://developer.intel.com/ipex-whl-stable-cpu
--extra-index-url https://download.pytorch.org/whl/cpu
torch==1.13.1
intel_extension_for_pytorch==1.13.100
accelerate==0.21.0
datasets==2.14.2
fsspec==2023.9.2
hydra-core==1.3.2
omegaconf==2.3.0
tiktoken==0.4.0
tqdm==4.65.0
transformers==4.30.0